+++ packages/app-content-pages/src/shared/stores/User/User.js
@@ -7,28 +7,11 @@ import numberString from '../types/numberString'
 
 import UserPersonalization from './UserPersonalization'
 
-async function decodeJWT(token) {
-  let user = null
-  let error = null
-  const decodedToken = await panoptesJS.auth.verify(token)
-  const { data } = decodedToken
-  error = decodedToken.error
-  if (data) {
-    user = {
-      id: data.id.toString(),
-      login: data.login,
-      display_name: data.dname,
-      admin: data.admin
-    }
-  }
-  return { user, error }
-}
-
 async function fetchPanoptesUser() {
   try {
     const token = await auth.checkBearerToken()
     if (token) {
-      const { user, error } = await decodeJWT(token)
+      const { user, error } = await panoptesJS.auth.decodeJWT(token)
       if (user) {
         return user
       }