+++ packages/lib-panoptes-js/docs/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.4.0] 2023-07-04
+- add `auth.decodeJWT(token)` to get a Panoptes user from a token.
+- allow `auth` methods to accept Authorization headers (which are used in the Classifier) as well as tokens.
+
 ## [0.3.0] 2023-05-25
 - add JWT verification as `auth.verify(token)`.